move Promise to package http
