+++ base/rico-core/src/main/java/dev/rico/core/http/Promise.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.rico.core.functional;
+package dev.rico.core.http;
 
 import java.util.concurrent.CompletableFuture;
 import java.util.function.Consumer;
